golang默认生成int64，不再使用int

+++ src/paladin/code_builder.go
@@ -188,7 +188,7 @@ func (p *CodeBuilder) AppendKeyword(code *jen.Statement, t reflect.Type) *jen.St
 		return code.Float64()
 
 	case reflect.Int:
-		return code.Int()
+		return code.Int64()
 
 	case reflect.String:
 		return code.String()
